feat: Add patch endpoint to CommentController
This commit adds a new endpoint `patch` to the `CommentController` class. The `patch` method allows updating a comment with the provided `id` using the `PatchCommentDto` in the request body. If the comment is found and successfully updated, a success message along with the updated comment data is returned. If the comment is not found, a failure message is returned. This functionality will be used to modify existing comments in the application.