feat: Add like and patch functionality to response controller
This commit adds two new endpoints to the `ResponseController` class. The `like` endpoint allows users to like or unlike a response by sending a `LikeResponseDto` in the request body. The `patch` endpoint allows updating a response with the provided `id` using the `PatchResponseDto` in the request body. If the response is found and successfully updated, a success message along with the updated response data is returned. If the response is not found, a failure message is returned. These functionalities will be used to handle liking and updating responses in the application.