refactor: Update CommentService timeline method to use query builder
This commit refactors the `timeline` method in the `CommentService` class to use the query builder instead of the `find` method. The new implementation joins the necessary tables (`sender`, `responses`, `responses.sender`, `responses.recipient`) and orders the results by the creation date of comments and responses in ascending order. This change improves the efficiency and readability of the code.