feat: Return created comment data in CommentController create method
This commit modifies the `create` method in the `CommentController` class to return the created comment data in the response. Previously, only a success message was returned. Now, after creating the comment using the `commentService.create` method, the created comment is assigned to a variable `comment` and included in the response data. This change provides more information to the client about the created comment.